Moving michaelis menten function to transformers module
As requested on issue #532
