+++ pymc_marketing/mmm/base.py
@@ -21,12 +21,14 @@ from sklearn.preprocessing import FunctionTransformer
 from xarray import DataArray, Dataset
 
 from pymc_marketing.mmm.budget_optimizer import budget_allocator
+from pymc_marketing.mmm.transformers import (
+    michaelis_menten,
+    sigmoid_saturation,
+)
 from pymc_marketing.mmm.utils import (
     estimate_menten_parameters,
     estimate_sigmoid_parameters,
-    extense_sigmoid,
     find_sigmoid_inflection_point,
-    michaelis_menten,
     standardize_scenarios_dict_keys,
 )
 from pymc_marketing.mmm.validating import (
@@ -526,7 +528,7 @@ class BaseMMM(ModelBuilder):
         # Estimate parameters based on the method
         if method == "sigmoid":
             estimate_function = estimate_sigmoid_parameters
-            fit_function = extense_sigmoid
+            fit_function = sigmoid_saturation
         elif method == "michaelis-menten":
             estimate_function = estimate_menten_parameters
             fit_function = michaelis_menten
@@ -778,7 +780,7 @@ class BaseMMM(ModelBuilder):
             x_inflection, y_inflection = find_sigmoid_inflection_point(
                 alpha=alpha_limit, lam=lam_constant
             )
-            fit_function = extense_sigmoid
+            fit_function = sigmoid_saturation
         elif method == "michaelis-menten":
             alpha_limit, lam_constant = estimate_menten_parameters(
                 channel=channel,
