+++ pymc_marketing/mmm/utils.py
@@ -7,6 +7,8 @@ import pandas as pd
 import xarray as xr
 from scipy.optimize import curve_fit, minimize_scalar
 
+from pymc_marketing.mmm.transformers import michaelis_menten, sigmoid_saturation
+
 
 def generate_fourier_modes(
     periods: npt.NDArray[np.float_], n_order: int
@@ -38,61 +40,6 @@ def generate_fourier_modes(
             for func in ("sin", "cos")
         }
     )
-
-
-def michaelis_menten(
-    x: Union[float, np.ndarray, npt.NDArray[np.float64]],
-    alpha: Union[float, np.ndarray, npt.NDArray[np.float64]],
-    lam: Union[float, np.ndarray, npt.NDArray[np.float64]],
-) -> Union[float, Any]:
-    """
-    Evaluate the Michaelis-Menten function for given values of x, alpha, and lambda.
-
-    The Michaelis-Menten function is a type of mathematical saturation function commonly used in
-    enzyme kinetics, but it's also applicable in marketing mix models to describe
-    how different channels contribute to a certain outcome (e.g., sales or conversions)
-    as the spending on that channel increases and the contribution saturates.
-
-    Mathematically, it is described as:
-    α * x / (λ + x)
-
-    Parameters
-    ----------
-    x : float
-        The spent on a channel.
-    alpha (Limit/Vmax) : float
-        The maximum contribution a channel can make.
-    lam (k) : float
-        The elbow on the function in `x` (Point where the curve change their direction).
-
-    Returns
-    -------
-    float
-        The value of the Michaelis-Menten function given the parameters.
-    """
-
-    return alpha * x / (lam + x)
-
-
-def extense_sigmoid(
-    x: Union[float, np.ndarray, npt.NDArray[np.float64]],
-    alpha: Union[float, np.ndarray, npt.NDArray[np.float64]],
-    lam: Union[float, np.ndarray, npt.NDArray[np.float64]],
-) -> Union[float, Any]:
-    """
-    Parameters
-    ----------
-    - alpha
-        α (alpha): Represent the Asymptotic Maximum or Ceiling Value.
-    - lam
-        λ (lambda): affects how quickly the function approaches its upper and lower asymptotes. A higher value of
-        lam makes the curve steeper, while a lower value makes it more gradual.
-    """
-
-    if alpha <= 0 or lam <= 0:
-        raise ValueError("alpha and lam must be greater than 0")
-
-    return (alpha - alpha * np.exp(-lam * x)) / (1 + np.exp(-lam * x))
 
 
 def estimate_menten_parameters(
@@ -180,7 +127,7 @@ def estimate_sigmoid_parameters(
 
     parameter_bounds_modified = ([0, 0], [alpha_initial_estimate, np.inf])
     popt, _ = curve_fit(
-        extense_sigmoid,
+        sigmoid_saturation,
         x,
         y,
         p0=[alpha_initial_estimate, lam_initial_estimate],
@@ -258,7 +205,7 @@ def find_sigmoid_inflection_point(
 
     # Evaluate the original function at the inflection point
     x_inflection = result.x
-    y_inflection = extense_sigmoid(x_inflection, alpha, lam)
+    y_inflection = sigmoid_saturation(x_inflection, alpha, lam)
 
     return x_inflection, y_inflection
